new define skeleton for puppet
The problem with this is defines in puppet aren't written as:
define_somethingneat.pp they are just somethingneat.pp and there's no
way to distinguish between starting a class and starting a define from
the filetype.  I plan to lazily avoid this with editing
define_something_neat.pp then saving the buffer to the right name and
having the vars drop the 'define_' from the front.