a few additions
added the ir_black colorscheme, trying to see if I like it I mostly do
now that I'm running in 256 colors it have more options, but I'm not
used to things so it's messing with me a bit.

some minor additions to puppet filetype to support better word jumping
and expanding of the word under cursor with 'K'

statusline updates so I know what things mean and add long lines, not
sure if the longlines I care about but in general most coding standards
don't want lines over 80 chars.