annoyances
adding <> to matchpairs causes a bell everytime you type one mismatched.
This is extremely annoying when they are often unmatched -> for ex.

nice little plugin to automatically create a directory if you vi a file
or open a buffer and the tree doesn't exist.  gisted straight from #vim,
now that is service.